deduplicate ip_name_lookup & relocate util helpers
Continuing the p2/p3 sockets unification of https://github.com/bytecodealliance/wasmtime/pull/13934, this PR deduplicates the `ip_name_lookup` implementations as well. And [as suggested](https://github.com/bytecodealliance/wasmtime/pull/13934#pullrequestreview-4757192591) gets rid of the `util.rs` file. Common helpers have been moved into `mod.rs`, and TCP/UDP-specific helpers have been moved into their respective files.
