+++ crates/wasi/src/p3/sockets/conv.rs
@@ -1,7 +1,8 @@
 use crate::p3::bindings::sockets::types;
 use crate::p3::sockets::SocketError;
-use crate::sockets::SocketAddressFamily;
-use crate::sockets::util::{from_ipv4_addr, from_ipv6_addr, to_ipv4_addr, to_ipv6_addr};
+use crate::sockets::{
+    SocketAddressFamily, from_ipv4_addr, from_ipv6_addr, to_ipv4_addr, to_ipv6_addr,
+};
 use core::net::{IpAddr, SocketAddr, SocketAddrV4, SocketAddrV6};
 use rustix::io::Errno;
 use std::net::ToSocketAddrs;
@@ -204,30 +205,49 @@ impl From<&Errno> for types::ErrorCode {
     }
 }
 
-impl From<crate::sockets::util::ErrorCode> for types::ErrorCode {
-    fn from(code: crate::sockets::util::ErrorCode) -> Self {
+impl From<crate::sockets::ErrorCode> for types::ErrorCode {
+    fn from(code: crate::sockets::ErrorCode) -> Self {
         match code {
-            crate::sockets::util::ErrorCode::Other => Self::Other(None),
-            crate::sockets::util::ErrorCode::AccessDenied => Self::AccessDenied,
-            crate::sockets::util::ErrorCode::NotSupported => Self::NotSupported,
-            crate::sockets::util::ErrorCode::InvalidArgument => Self::InvalidArgument,
-            crate::sockets::util::ErrorCode::OutOfMemory => Self::OutOfMemory,
-            crate::sockets::util::ErrorCode::Timeout => Self::Timeout,
-            crate::sockets::util::ErrorCode::InvalidState => Self::InvalidState,
-            crate::sockets::util::ErrorCode::AddressNotBindable => Self::AddressNotBindable,
-            crate::sockets::util::ErrorCode::AddressInUse => Self::AddressInUse,
-            crate::sockets::util::ErrorCode::RemoteUnreachable => Self::RemoteUnreachable,
-            crate::sockets::util::ErrorCode::ConnectionRefused => Self::ConnectionRefused,
-            crate::sockets::util::ErrorCode::ConnectionBroken => Self::ConnectionBroken,
-            crate::sockets::util::ErrorCode::ConnectionReset => Self::ConnectionReset,
-            crate::sockets::util::ErrorCode::ConnectionAborted => Self::ConnectionAborted,
-            crate::sockets::util::ErrorCode::DatagramTooLarge => Self::DatagramTooLarge,
+            crate::sockets::ErrorCode::Other => Self::Other(None),
+            crate::sockets::ErrorCode::AccessDenied => Self::AccessDenied,
+            crate::sockets::ErrorCode::NotSupported => Self::NotSupported,
+            crate::sockets::ErrorCode::InvalidArgument => Self::InvalidArgument,
+            crate::sockets::ErrorCode::OutOfMemory => Self::OutOfMemory,
+            crate::sockets::ErrorCode::Timeout => Self::Timeout,
+            crate::sockets::ErrorCode::InvalidState => Self::InvalidState,
+            crate::sockets::ErrorCode::AddressNotBindable => Self::AddressNotBindable,
+            crate::sockets::ErrorCode::AddressInUse => Self::AddressInUse,
+            crate::sockets::ErrorCode::RemoteUnreachable => Self::RemoteUnreachable,
+            crate::sockets::ErrorCode::ConnectionRefused => Self::ConnectionRefused,
+            crate::sockets::ErrorCode::ConnectionBroken => Self::ConnectionBroken,
+            crate::sockets::ErrorCode::ConnectionReset => Self::ConnectionReset,
+            crate::sockets::ErrorCode::ConnectionAborted => Self::ConnectionAborted,
+            crate::sockets::ErrorCode::DatagramTooLarge => Self::DatagramTooLarge,
         }
     }
 }
 
-impl From<crate::sockets::util::ErrorCode> for SocketError {
-    fn from(code: crate::sockets::util::ErrorCode) -> Self {
+impl From<crate::sockets::ErrorCode> for SocketError {
+    fn from(code: crate::sockets::ErrorCode) -> Self {
         SocketError::from(types::ErrorCode::from(code))
+    }
+}
+
+impl From<crate::sockets::ip_name_lookup::ErrorCode>
+    for crate::p3::bindings::sockets::ip_name_lookup::ErrorCode
+{
+    fn from(code: crate::sockets::ip_name_lookup::ErrorCode) -> Self {
+        match code {
+            crate::sockets::ip_name_lookup::ErrorCode::AccessDenied => Self::AccessDenied,
+            crate::sockets::ip_name_lookup::ErrorCode::InvalidArgument => Self::InvalidArgument,
+            crate::sockets::ip_name_lookup::ErrorCode::NameUnresolvable => Self::NameUnresolvable,
+            crate::sockets::ip_name_lookup::ErrorCode::TemporaryResolverFailure => {
+                Self::TemporaryResolverFailure
+            }
+            crate::sockets::ip_name_lookup::ErrorCode::PermanentResolverFailure => {
+                Self::PermanentResolverFailure
+            }
+            crate::sockets::ip_name_lookup::ErrorCode::Other => Self::Other(None),
+        }
     }
 }
