+++ crates/wasi/src/p3/sockets/host/ip_name_lookup.rs
@@ -1,38 +1,20 @@
-use tokio::net::lookup_host;
-use wasmtime::component::Accessor;
-
 use crate::p3::bindings::sockets::ip_name_lookup::{ErrorCode, Host, HostWithStore};
 use crate::p3::bindings::sockets::types;
 use crate::p3::sockets::WasiSockets;
 use crate::sockets::WasiSocketsCtxView;
-use crate::sockets::util::{from_ipv4_addr, from_ipv6_addr, parse_host};
+use crate::sockets::ip_name_lookup::resolve_addresses;
+use wasmtime::component::Accessor;
 
 impl<U> HostWithStore<U> for WasiSockets {
     async fn resolve_addresses(
         store: &Accessor<U, Self>,
         name: String,
     ) -> wasmtime::Result<Result<Vec<types::IpAddress>, ErrorCode>> {
-        let Ok(host) = parse_host(&name) else {
-            return Ok(Err(ErrorCode::InvalidArgument));
-        };
-        if !store.with(|mut view| view.get().ctx.allowed_network_uses.ip_name_lookup) {
-            return Ok(Err(ErrorCode::PermanentResolverFailure));
-        }
-        match host {
-            url::Host::Ipv4(addr) => Ok(Ok(vec![types::IpAddress::Ipv4(from_ipv4_addr(addr))])),
-            url::Host::Ipv6(addr) => Ok(Ok(vec![types::IpAddress::Ipv6(from_ipv6_addr(addr))])),
-            url::Host::Domain(domain) => {
-                // This is only resolving names, not ports, so force the port to be 0.
-                if let Ok(addrs) = lookup_host((domain.as_str(), 0)).await {
-                    Ok(Ok(addrs
-                        .map(|addr| addr.ip().to_canonical().into())
-                        .collect()))
-                } else {
-                    // If/when we use `getaddrinfo` directly, map the error properly.
-                    Ok(Err(ErrorCode::NameUnresolvable))
-                }
-            }
-        }
+        let fut = store.with(|mut view| resolve_addresses(&view.get().ctx, name));
+        Ok(match fut.await {
+            Ok(addrs) => Ok(addrs.into_iter().map(|addr| addr.into()).collect()),
+            Err(err) => Err(err.into()),
+        })
     }
 }
